Added Centipede, mostly working
Added Magnitude function for joystick so there is a build up on the velocity of the joystick.

Added Laser as a project, with collision detection.

Added Centipede,

Added Collision detection, and main code to spawn mushrooms, centipede,

Resets when Centipede collides with player

Resets when centipede is fully destroyed.

Added Scoreboard